backend/requirements.txt
amqp==5.2.0
ansicon==1.89.0
APScheduler==3.10.4
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.7
daphne==3.0.2
Django==4.2.13
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-picklefield==3.2
django-redis==5.4.0
djangorestframework==3.15.1
dnspython==2.6.1
eventlet==0.36.1
exceptiongroup==1.2.1
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jinxed==1.2.1
kombu==5.3.7
msgpack==1.0.8
msgpack-python==0.5.6
outcome==1.3.0.post0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
requests==2.32.2
selenium==4.21.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
tornado==6.4
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
zope.interface==6.4.post2
uwsgi
